Add set.ts and fix pipe fluent api
