+++ src/pipe.ts
@@ -8,9 +8,17 @@
      ## ## ## :##
       ## ## ##*/
 
-import { set, AccessorFunction, ValueTransformer } from '.'
+import { set, AccessorFunction, ValueTransformer } from './set'
 
-export function pipe<R>(input: R) {
+export type Pipe<R> = {
+  set: <T, A extends AccessorFunction<R, T>>(
+    accessor: A,
+    value: ValueTransformer<A>
+  ) => Pipe<R>
+  end: () => R
+}
+
+export function pipe<R>(input: R): Pipe<R> {
   return {
     set: <T, A extends AccessorFunction<R, T>>(
       accessor: A,
